+++ src/Blogr.php
@@ -4,5 +4,5 @@ namespace Happytodev\BlogrDocs;
 
 class Blogr
 {
-    const VERSION = '1.5.0';
+    const VERSION = '1.6.0';
 }